rename the Token class to SessionToken
Emphasises that it only handles the kind of token that's bound to a session -
as opposed to registration tokens or password recovery tokens.

Sadly, this makes some lines in oauth.php kinda long.
I'll probably mitigate this by finally creating a new method for "branching
off" a token for the same service.