assign random legacy_ids to invited users
Kinda conflicted about this.
This introduces a new confusing failure mode (future admins won't know what to
do if there's a legacy_id conflict). The chances of someone not being assigned
a legacy_id are low enough that, if you have some people with a NULL legacy_id,
you might not notice it when migrating the data away.