elder.php: rearrange the columns, visual tweaks
I've arranged the columns in a way that I think makes more sense.
The legacy IDs are now monospace because they obviously should be monospace
(they're random 20 character strings). I'm also hiding the column that has
them because otherwise this site is unusable on mobile, but if you're on
mobile then you probably don't give a fuck about the ID.